+++ base/models/teaching_material.py
@@ -29,8 +29,6 @@ from ordered_model.admin import OrderedModelAdmin
 from ordered_model.models import OrderedModel
 from reversion.admin import VersionAdmin
 
-from base.business.learning_units.pedagogy import update_bibliography_changed_field_in_cms
-
 
 class TeachingMaterialAdmin(VersionAdmin, OrderedModelAdmin):
     list_display = ('title', 'mandatory', 'learning_unit_year', 'order', 'move_up_down_links')
@@ -56,28 +54,3 @@ class TeachingMaterial(OrderedModel):
 def find_by_learning_unit_year(learning_unit_year):
     return TeachingMaterial.objects.filter(learning_unit_year=learning_unit_year)\
                                    .order_by('order')
-
-
-def postpone_teaching_materials(start_luy, commit=True):
-    """
-    This function override all teaching materials from start_luy until latest version of this luy
-    :param start_luy: The learning unit year which we want to start postponement
-    :param commit:
-    :return:
-    """
-    teaching_materials = find_by_learning_unit_year(start_luy)
-    for next_luy in [luy for luy in start_luy.find_gt_learning_units_year()]:
-        # Remove all previous teaching materials
-        next_luy.teachingmaterial_set.all().delete()
-        # Inserts all teaching materials comes from start_luy
-        to_inserts = [TeachingMaterial(title=tm.title, mandatory=tm.mandatory, learning_unit_year=next_luy)
-                      for tm in teaching_materials]
-        bulk_save(to_inserts, commit)
-
-        # For sync purpose, we need to trigger an update of the bibliography when we update teaching materials
-        update_bibliography_changed_field_in_cms(next_luy)
-
-
-def bulk_save(teaching_materials, commit=True):
-    for teaching_material in teaching_materials:
-        teaching_material.save(commit)
