+++ base/tests/models/test_teaching_material.py
@@ -25,7 +25,8 @@
 ##############################################################################
 from django.test import TestCase
 
-from base.models.teaching_material import postpone_teaching_materials, TeachingMaterial, find_by_learning_unit_year
+from base.models.teaching_material import TeachingMaterial, find_by_learning_unit_year
+from base.business.learning_units.pedagogy import postpone_teaching_materials
 from base.tests.factories.business.learning_units import GenerateAcademicYear
 from base.tests.factories.learning_container import LearningContainerFactory
 from base.tests.factories.learning_unit import LearningUnitFactory
